Move predefined surfaces to 'surfaces' dir
