+++ src/CMakeLists.txt
@@ -12,4 +12,5 @@ endforeach()
 list(REMOVE_DUPLICATES dir_list)
 include_directories(${dir_list})
 
-file(COPY gepard.h gepard-surface.h gepard-xsurface.h gepard-memory-buffer-surface.h DESTINATION ${INCLUDE_OUTPUT_DIR})
+file(COPY gepard.h DESTINATION ${INCLUDE_OUTPUT_DIR})
+file(COPY surfaces DESTINATION ${INCLUDE_OUTPUT_DIR})
