+++ src/engines/software/gepard-software.h
@@ -30,8 +30,8 @@
 
 #include "gepard-defs.h"
 
+#include "gepard.h"
 #include "gepard-image.h"
-#include "gepard-surface.h"
 #include "gepard-types.h"
 
 namespace gepard {
